add test on build script

+++ build.sh
@@ -2,6 +2,18 @@
 
 ARCHIVE_NAME="voiting_game.tar.gz"
 
+# Перехід до каталогу backend і запуск тестів
+cd backend
+npm test
+
+# Якщо тести не пройшли, зупиняємо скрипт
+if [ $? -ne 0 ]; then
+    echo "Тести не пройшли! Зупиняємо білд."
+    exit 1
+fi
+
+cd ..
+
 # Видаляємо старий архів, якщо він існує
 rm -f $ARCHIVE_NAME
